+++ app/src/main/AndroidManifest.xml
@@ -13,35 +13,35 @@
         android:theme="@style/Theme.LAMASYA"
         tools:targetApi="31">
         <activity
-            android:name=".ui.login.ResetPasswordActivity"
+            android:name=".ui.view.login.ResetPasswordActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".ui.register.RegisterActivity"
+            android:name=".ui.view.register.RegisterActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".ui.login.LoginActivity"
+            android:name=".ui.view.login.LoginActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".ui.main.MainActivity"
+            android:name=".ui.view.main.MainActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".ui.onBoaring.OnBoardingActivity"
+            android:name=".ui.view.onBoaring.OnBoardingActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
