+++ app/src/main/java/com/lamasya/ui/viewmodel/ResetPwViewModel.kt
@@ -1,4 +1,4 @@
-package com.lamasya.ui.login
+package com.lamasya.ui.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
