+++ app/src/main/res/layout/activity_reset_password.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.login.ResetPasswordActivity">
+    tools:context=".ui.view.login.ResetPasswordActivity">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/EmailTextInput"
