+++ app/src/main/res/values/strings.xml
@@ -7,6 +7,7 @@
     <string name="tittle_notification">Notification</string>
 
     <!--    material-->
+    <string name="welcome_to_lamasya">Welcome to Lamasya</string>
     <string name="last_name">Last Name</string>
     <string name="first_name">First Name</string>
     <string name="email">Email</string>
